FastAPI-Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-google-genai
langchain-pinecone
pinecone
sentence-transformers
